+++ app/src/main/java/com/example/e2tech/Models/VoucherModel.java
@@ -6,13 +6,13 @@ import java.io.Serializable;
 
 public class VoucherModel extends ViewModel implements Serializable {
     String code;
-    double discount;
+    int discount;
     String description;
     String img_url;
 
     public VoucherModel() {}
 
-    public VoucherModel(String code, double discount, String description, String img_url) {
+    public VoucherModel(String code, int discount, String description, String img_url) {
         this.code = code;
         this.discount = discount;
         this.description = description;
@@ -27,11 +27,11 @@ public class VoucherModel extends ViewModel implements Serializable {
         this.code = code;
     }
 
-    public double getDiscount() {
+    public int getDiscount() {
         return discount;
     }
 
-    public void setDiscount(double discount) {
+    public void setDiscount(int discount) {
         this.discount = discount;
     }
